x86/iR5900: Fix signed/unsigned comparison in timeout loop

+++ pcsx2/x86/ix86-32/iR5900.cpp
@@ -2353,7 +2353,7 @@ static void recRecompile(const u32 startpc)
 			else if ((cpuRegs.code >> 26) == 5)
 			{
 				// bne
-				if (timeout_reg != _Rs_ || _Rt_ != 0 || memRead32(i + 4) != 0)
+				if (timeout_reg != static_cast<s32>(_Rs_) || _Rt_ != 0 || memRead32(i + 4) != 0)
 					is_timeout_loop = false;
 			}
 			else if (cpuRegs.code != 0)
